+++ samples/boards/nrf52/power_mgr/prj_tickless.conf
@@ -1,15 +1,9 @@
-CONFIG_ASSERT=y
-CONFIG_ASSERT_LEVEL=2
 CONFIG_NUM_COOP_PRIORITIES=29
 CONFIG_NUM_PREEMPT_PRIORITIES=40
-
-CONFIG_SYS_POWER_MANAGEMENT=y
-CONFIG_TICKLESS_IDLE=y
-CONFIG_TICKLESS_KERNEL=y
-
 CONFIG_SYS_POWER_MANAGEMENT=y
 CONFIG_SYS_POWER_LOW_POWER_STATE=y
 CONFIG_SYS_POWER_DEEP_SLEEP=y
 CONFIG_DEVICE_POWER_MANAGEMENT=y
+CONFIG_TICKLESS_KERNEL=y
 CONFIG_TICKLESS_IDLE=y
 CONFIG_GPIO=y
